+++ src/main/java/com/hubspot/jinjava/lib/filter/FilterLibrary.java
@@ -1,17 +1,17 @@
 /**********************************************************************
-Copyright (c) 2014 HubSpot Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+ * Copyright (c) 2014 HubSpot Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  **********************************************************************/
 package com.hubspot.jinjava.lib.filter;
 
@@ -29,12 +29,12 @@ public class FilterLibrary extends SimpleLibrary<Filter> {
         AttrFilter.class,
         PrettyPrintFilter.class,
 
-        DefaultFilter.class,
+    DefaultFilter.class,
         DAliasedDefaultFilter.class,
         FileSizeFormatFilter.class,
         UrlizeFilter.class,
 
-        BatchFilter.class,
+    BatchFilter.class,
         CountFilter.class,
         DictSortFilter.class,
         FirstFilter.class,
@@ -52,11 +52,12 @@ public class FilterLibrary extends SimpleLibrary<Filter> {
         ShuffleFilter.class,
         SortFilter.class,
         SplitFilter.class,
+        UniqueFilter.class,
 
-        DatetimeFilter.class,
+    DatetimeFilter.class,
         DateTimeFormatFilter.class,
 
-        AbsFilter.class,
+    AbsFilter.class,
         AddFilter.class,
         CutFilter.class,
         DivideFilter.class,
@@ -70,14 +71,14 @@ public class FilterLibrary extends SimpleLibrary<Filter> {
         RoundFilter.class,
         SumFilter.class,
 
-        EscapeFilter.class,
+    EscapeFilter.class,
         EAliasedEscapeFilter.class,
         ForceEscapeFilter.class,
         StripTagsFilter.class,
         UrlEncodeFilter.class,
         XmlAttrFilter.class,
 
-        CapitalizeFilter.class,
+    CapitalizeFilter.class,
         CenterFilter.class,
         FormatFilter.class,
         IndentFilter.class,